Update IntegralScheme parsing to use integer type; remove Gauss integration module as it is no longer needed; refactor main execution to utilize MesMatrixElements for matrix calculations and adjust test input files for consistency.

+++ file_parse.py
@@ -31,7 +31,7 @@ class GlobalData:
                 continue
 
             if line.startswith("IntegralScheme"):
-                self.N = float(line.split()[1])
+                self.N = int(line.split()[1])
             elif line.startswith("SimulationTime"):
                 self.SimulationTime = float(line.split()[1])
             elif line.startswith("SimulationStepTime"):
